CCNET-1019: DevEnv task misinterprets spaces in project path
The <devenv> task has been changed to enquote the solution file name, the configuration name, and the project name.  Since some people may have been working around this by enquoting them in the CCNet configuration, the fix is conditional - it doesn't add quotes if the value starts with a quote.