CCNET-887: Hopeful fix for the overcaching. Stopped rendering the NOT_AVAILABLE fingerprint to the response headers. It appears this was leading to endless caching, because the client and server both agreed that the only fingerprint for the content was the NOT_AVAILABLE fingerprint.
This fix is less aggressive than what I was doing at first: it may still be possible for clients to send bad request headers such that they get a NotModified response when they shouldn't, but that's their fault. NOT_AVAILABLE should not be equal to itself anyway, and behaves that way in unit tests, so I'm not clear EXACTLY what was happening.