First cut of an encrypted communications channel. This will encrypt messages over HTTP and .NET Remoting using Rijndael encryption.
The process uses the following pattern:
1) On the first call, retrieve the public key from the server.
2) Generate a new secret key, encrypt it and pass it back to the server - the server then stores this key.

On subquent calls use the secret key to encrypt all messages.

This is done by adding a new communications channel called EncryptingConnection. This channel acts as a wrapper over top of the original channel (e.g. HTTP or .NET remoting) and encrypts/decrypts the messages that pass through.

Encryption is set on the client side, although the server can be configured to enforce encryption.