CCNET-1911: Svn: TagOnSuccess TagWorkingCopy - save version number generated by ccnetLabel when tagged
This patch adds an optional flag TagWorkingCopy to the Svn source
control block. If an AssemblyInfo.cs or CommonAssemblyInfo.cs has been
modified (as in autoversioning), those modifications will be saved
when when the tag is created.

When TagWorkingCopy is true the tagBase will be the working copy,
and the --revision xx will be omitted.

The patch includes a unit test
CreateLabelFromWorkingCopyWhenTagWorkingCopyTrue that tests the
TagWorkingCopy flag.

Patch by Ron Vander Wal