Fix: Nasty cross session bug: one session could view the project list with the rights of another user
-> The SessionToken of the last freshly logged in user that made a round-trip to the server is used for filtering the project list.

Not sure if this is sufficient to completely solve the issue. There might be an race issue left due to client caching, but I don't know if requests can be handled in parallel for several sessions. If yes, then caching shall be disabled or the SessionToken property removed from the client.