Fix: the session token was not exactly stored in what is called a "session cookie"
The cookie had 15min expiration, a session cookie has no Expires property. Now the cookie is a true session cookie and lasts for the whole session. Of course the session token used is checked by the server so the duration of validity of the token shall be adjusted in the server configuration to meet the intended goal.