Fix: Unhandled NoSuchProjectException yielding user unfriendly exception page
If the user gets automatically logged out (session timeout) from the server, a NoSuchProjectException exception is triggered on the call to GetCategory(). This generates an error page which is irritating for users. We mask the exception, a subsequent remoting call is handled already and generates an appropriate message in the dashboard: "The session token is either invalid or is for a session that has expired."