Fixes bug #39 - packageFolder heirarchy problem
The problem was that the resulting file path was always the combination
of the targetFolder and filename. When subdirectories were involved,
this produced problems if you wanted the generated zip to respect
hierarchy. Fixed the issue by recursively processing sub folders
individually versus using the DirectoryInfo.GetFiles() with the
SearchOption.AllDirectories option.