Add https support to CCTray
This change adds security protocols to the CruiseServerHttpClient which allows the application to handle build servers that communicate over "https". (e.g. jenkins with cc.xml)
Since there is currently an issue with the jenkins cc.xml returning "http"-urls for build jobs while the server is running over "https", this change also contains a workaround to handle this issue by changing "http" to "https" in case the build server is using "https".